+++ src/all.ts
@@ -12,6 +12,7 @@ export { CpsChip } from './components/chip.js';
 export { CpsDropdown } from './components/dropdown.js';
 export { CpsFlyout } from './components/flyout.js';
 export { CpsFormatDate } from './components/format-date.js';
+export { CpsFormatNumber } from './components/format-number.js';
 export { CpsIcon } from './components/icon.js';
 export { CpsIconButton } from './components/icon-button.js';
 export { CpsInclude } from './components/include.js';